Add USB HID gadget driver fixes for XAC
Hack the gadget HID driver so it works with the XAC. For more details see
rpi-5.10/README.md.

8_buttons_joystick_usb change to make the USB descriptors closer to
a real joystick. The whole script must be run as root so the extra sudo
commands have been removed.

8_Buttons_Joystick/joystick_inputevent.py gets BT keyboard and mouse input events.
Very minimal support but the left mouse button is mapped to Xbox A button and
right mouse button is mapped to Xbox B button.